Correzioni e miglioramenti vari nei controller
- Corretto errore di chiusura del namespace in `CartController.cs` e `EventController.cs`.
- Modificata la logica del metodo `DeleteDj` in `DjController.cs`:
  - Rinominato in `Delete` con parametro `confirm` opzionale.
  - Gestione della conferma per eliminazione DJ.
- Aggiunto caricamento generi musicali nella lista eventi in `EventController.cs`.
- Aggiunto caricamento immagini eventi e location in `TicketController.cs`.
- Aggiunta funzione `GetPopularEventsAsync` in `EventService.cs` per ottenere eventi popolari basati sui commenti recenti.

+++ Capstone/Controllers/DjController.cs
@@ -54,11 +54,18 @@ public class DjController : Controller
         return View(djList);
     }
 
-    [HttpPost]
-    public async Task<IActionResult> DeleteDj(int id)
+    // GET: Dj/Delete/5
+    public async Task<IActionResult> Delete(int id, bool confirm = false)
     {
-        await _djSvc.DeleteDjAsync(id);
-        return RedirectToAction("List");
+        if (confirm)
+        {
+            await _djSvc.DeleteDjAsync(id);
+            return RedirectToAction("List");
+        }
+
+        var djToDelete = await _djSvc.GetDjByIdAsync(id);
+
+        return View(djToDelete);
     }
 
     public async Task<IActionResult> Edit(int id)
